Priority Nav: Trigger on all challenge pages
Moves the P-nav code from Pamola implementation to stimulusjs
to ensure that the code runs on all challenge pages.
