+++ app/assets/stylesheets/application-redesign.css.scss
@@ -12079,9 +12079,6 @@ body {
 
 @media (min-width: 0px) and (max-width: 479px) {
   .sub-nav-bar .nav-item {
-    margin: 0;
-  }
-  .sub-nav-bar .nav-item:first-of-type {
     margin-right: 1.5rem;
   }
 }
