+++ app/views/challenges/show/_subnav.html.erb
@@ -1,6 +1,6 @@
 <!-- sub-nav bar -->
 <div class="sub-nav-bar sticky-top">
-  <div class="container-fluid priority-nav">
+  <div class="container-fluid priority-nav" data-controller="prioritynav">
     <ul class="nav">
       <%= render partial: 'challenges/show/subnav_link', locals: {text: 'Overview', link: challenge_path(challenge)} %>
       <%= render partial: 'challenges/show/subnav_link', locals: {text: 'Discussion', link: discourse_url(challenge)} %>
@@ -29,13 +29,13 @@
         <%= render partial: 'challenges/show/subnav_link', locals: {text: 'Rules', link: [challenge, challenge.current_challenge_rules]} %>
       <% end %>
 
-      <li class="overflow-nav nav-item dropdown">
+      <li class="overflow-nav nav-item dropdown ml-auto">
         <a class="nav-link overflow-nav-title" data-toggle="dropdown" href="#" role="button">
           <svg width="16" height="4" xmlns="http://www.w3.org/2000/svg">
             <path d="M2 0C.9 0 0 .9 0 2s.9 2 2 2 2-.9 2-2-.9-2-2-2zm12 0c-1.1 0-2 .9-2 2s.9 2 2 2 2-.9 2-2-.9-2-2-2zM8 0C6.9 0 6 .9 6 2s.9 2 2 2 2-.9 2-2-.9-2-2-2z" fill="#5D5F60" fill-rule="nonzero"/>
           </svg>
         </a>
-        <ul class="overflow-nav-list dropdown-menu"></ul>
+        <ul class="overflow-nav-list dropdown-menu dropdown-menu-right"></ul>
       </li>
     </ul>
